+++ irs-ess/src/main/java/org/eclipse/tractusx/ess/controller/EssRecursiveController.java
@@ -27,9 +27,9 @@ import io.swagger.v3.oas.annotations.Hidden;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.tractusx.irs.edc.client.exceptions.EdcClientException;
-import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
 import org.eclipse.tractusx.ess.service.EssRecursiveService;
+import org.eclipse.tractusx.irs.edc.client.model.notification.EdcNotification;
+import org.eclipse.tractusx.irs.edc.client.model.notification.InvestigationNotificationContent;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -53,8 +53,8 @@ public class EssRecursiveController {
 
     @PostMapping("/receive-recursive")
     @ResponseStatus(HttpStatus.CREATED)
-    public void registerRecursiveBPNInvestigation(final @Valid @RequestBody EdcNotification notification) throws
-            EdcClientException {
+    public void registerRecursiveBPNInvestigation(
+            final @Valid @RequestBody EdcNotification<InvestigationNotificationContent> notification) {
         log.info("Received recursive notification, starting investigation.");
         essRecursiveService.handleNotification(notification);
     }