+++ irs-ess/src/main/java/org/eclipse/tractusx/ess/service/EdcRegistration.java
@@ -152,8 +152,10 @@ public class EdcRegistration {
                               "odrl:or": [
                                 {
                                   "@type": "Constraint",
-                                  "odrl:leftOperand": "idsc:PURPOSE",
-                                  "odrl:operator": "EQ",
+                                  "odrl:leftOperand": "PURPOSE",
+                                  "odrl:operator": {
+                                    "@id": "odrl:eq"
+                                  },
                                   "odrl:rightOperand": "ID 3.0 Trace"
                                 }
                               ]