+++ irs-ess/src/test/java/org/eclipse/tractusx/ess/controller/EssControllerTest.java
@@ -112,7 +112,7 @@ class EssControllerTest {
                                                                          .globalAssetId(globalAssetId)
                                                                          .bpn(bpn)
                                                                          .build())
-                                          .incidentBpns(bpns)
+                                          .incidentBPNSs(bpns)
                                           .build();
     }