+++ irs-models/src/main/java/org/eclipse/tractusx/irs/component/partsiteinformationasplanned/Site.java
@@ -21,38 +21,18 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.component;
+package org.eclipse.tractusx.irs.component.partsiteinformationasplanned;
 
-import com.fasterxml.jackson.annotation.JsonUnwrapped;
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
+import java.time.ZonedDateTime;
+
 import lombok.Builder;
-import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
 
 /**
- * Relationship
+ * Site as part of {@link PartSiteInformationAsPlanned} according to semantic model version 1.0.0.
  */
-@Value
+@Builder
 @Jacksonized
-@Builder(toBuilder = true)
-@AllArgsConstructor
-@Schema(description = "Relationships between parent and child items.")
-public class Relationship {
-
-    private static final int GLOBAL_ASSET_ID_LENGTH = 45;
-
-    @Schema(implementation = String.class, description = "CATENA-X global asset id in the format urn:uuid:uuid4.",
-            example = "urn:uuid:6c311d29-5753-46d4-b32c-19b918ea93b0", minLength = GLOBAL_ASSET_ID_LENGTH,
-            maxLength = GLOBAL_ASSET_ID_LENGTH,
-            pattern = "^urn:uuid:[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$")
-    @JsonUnwrapped
-    private GlobalAssetIdentification catenaXId;
-
-    private LinkedItem linkedItem;
-
-    private String aspectType;
-
-    private String bpn;
-
+public record Site(ZonedDateTime functionValidUntil, String function, ZonedDateTime functionValidFrom,
+                   String catenaXSiteId) {
 }